add i18n

+++ src/main/java/com/project/shopapp/service/OrderDetailService.java
@@ -62,7 +62,11 @@ public class OrderDetailService implements IOrderDetailService {
 
     @Override
     public void deleteOrderDetail(Long id) {
-        orderDetailRepository.deleteById(id);
+        OrderDetail order = orderDetailRepository.findById(id).orElseThrow(null);
+        if (order != null){
+            orderDetailRepository.deleteById(id);
+        }
+
     }
 
     @Override
